fix 和 添加 tab 补全

+++ wsl2-debian.bash
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-# 如果存在 /etc/apt/sources.list.bak 跳过
+# 定义标志文件路径
+SCRIPT_FLAG="/var/log/.wsl2_debian_bash"
+
+# 检查标志文件是否存在
+if [ -f "$SCRIPT_FLAG" ]; then
+    echo "脚本已经执行过，跳过执行。"
+    exit 0
+fi
+
+# 如果存在 /etc/apt/sources.list.bak 跳过 , 避免多次覆盖 /etc/apt/sources.list.bak
 if [ ! -f /etc/apt/sources.list.bak ]; then
     # 备份原地址
     sudo cp  /etc/apt/sources.list /etc/apt/sources.list.bak
@@ -9,8 +18,6 @@ if [ ! -f /etc/apt/sources.list.bak ]; then
     sudo sed -i "s@http://deb.debian.org/debian@http://mirrors.tencent.com/debian@g" /etc/apt/sources.list
 
 fi
-
-
 
 sudo apt update
 
@@ -37,11 +44,33 @@ sudo apt install -y man-db manpages-zh
 # 其他必要工具
 sudo apt install -y curl wget xz-utils
 
-# 快速访问主机的下载文件目录 , jsw 为主机用户名
-sudo ln -s /mnt/c/Users/jsw/Downloads/ host_download
+# 快速访问主机的下载文件目录 , jsw 为主机用户名, 换成自己的
+sudo ln -s /mnt/c/Users/jsw/Downloads/ ~/host_download
 
 # 编译 内核
 sudo apt install -y gcc make bc libssl-dev libelf-dev flex bison libncurses-dev
 
+# 个人配置
+# 命令提示符 换个行
+sudo echo 'export PS1="$PS1\n\\$ "' | sudo tee -a ~/.bashrc > /dev/null
+
+# TAB 补全
+sudo apt install -y bash-completion
+
+# 配置 root 也可以 tab 键补全 , 内容是从 /etc/bash.bashrc 复制的
+sudo echo '# enable bash completion in interactive shells' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo 'if ! shopt -oq posix; then' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo '  if [ -f /usr/share/bash-completion/bash_completion ]; then' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo '    . /usr/share/bash-completion/bash_completion' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo '  elif [ -f /etc/bash_completion ]; then' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo '    . /etc/bash_completion' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo '  fi' | sudo tee -a /etc/bash.bashrc > /dev/null
+sudo echo 'fi' | sudo tee -a /etc/bash.bashrc > /dev/null
+
+
 # 设置中文
 sudo dpkg-reconfigure locales
+
+# 创建标志文件，表示脚本已执行
+sudo touch "$SCRIPT_FLAG"
+echo "脚本执行完成，已创建标志文件 $SCRIPT_FLAG"
